Bugfix: fix shadow shimmering
- Fix shadow shimmering artefact in issue #5
There is shadow shimmering artefact as describe in issue #5. This artefact is caused by the size changes on each cascade every time. If we look at the code we will find that the radius of the cascade is calculated every frame and is dependent on the orientation and position of camera. Because of floating point precision limitation, the radius changed slightly every frames if the camera move or rotate. We fix this issue by calculating the radius without using camera orientation and position.
- Add more info in soul_assert at GLExt::ProgramCreate function

Resolves #5