refactor: allow proxy allocator to accept parameter `size = 0`.
Some allocator doesn't need size when deallocating the memroy, but some proxy like ClearValueProxy and BoundGuardProxy need size to do its job.  In this changes we allow ProxyAllocator to accept `size = 0` and `addr = nullptr`, which mean that the callee don't know the size of the allocation. We add assert on some proxy that need size and allow them to accept nullptr. If addr is nullptr, proxy will do nothing.