Move resource ownership info into ResourceCacheState
- This changes will modify BinarySemaphore to include state, thus now command_queue will accept pointer to BinarySemaphore.
The reason for this is becase ResourceCacheState only store QueueType, thus we will need to check the BinarySemaphore state to see whether swapchain texture is owned by presentation engine.
- This changes also add sync_stages to ResourceCacheState.
Any previous access that has been synchronized is synchronized to to sync_stages, so only access in those stages are valid to avoid any hazard.